chore: reorganise-insight-scene-to-lift-higher
